python_requirements.txt
biopython==1.65
cycler==0.10.0
ipykernel==4.5.2
ipython==4.0.0
ipython-genutils==0.1.0
ipywidgets==4.0.3
Jinja2==2.9.4
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.4.0
jupyter-console==4.0.0
jupyter-contrib-core==0.3.0
jupyter-contrib-nbextensions==0.2.4
jupyter-core==4.2.1
jupyter-highlight-selected-word==0.0.8
jupyter-latex-envs==1.3.7.2
matplotlib==2.0.0
nbconvert==5.0.0
nbformat==4.2.0
networkx==1.11
nose==1.3.7
notebook==4.3.1
numpy==1.12.0
pandas==0.19.1
pyfaidx==0.4.2
PyYAML==3.11
pyzmq==16.0.2
scikit-learn==0.18.1
scipy==0.18.1
seaborn==0.8.dev0
tornado==4.4.2
